+++ npcs/base.npctype.patch
@@ -101,11 +101,5 @@
       "deltaValue" : 0.0,
       "initialPercentage" : 0.0
     }
-  },
-
-  {
-    "op" : "add",
-    "path" : "/statusControllerSettings/primaryScriptSources/-",
-    "value" : "/stats/sexbound/npc_primary.lua"
   }
 ]
