+++ scripts/sexbound/lib/sexbound/actor.lua
@@ -643,7 +643,7 @@ function Sexbound.Actor:setup(actorConfig)
     if self._config.inHeat then actorStatus:addStatus("sexbound_aroused_heat") end
     if self._config.isDefeated then actorStatus:addStatus("sexbound_defeated") end
     if self._config.birthcontrol then actorStatus:addStatus("birthcontrol") end
-    if self._config.canUseSexUIDefeated then actorStatus:addStatus("sexbound_defeated_can_use_ui") end
+    if self._config.canUseSexUIDefeated then actorStatus:addStatus("can_use_sex_ui_defeated") end
     
     if self._config.identity.sxbNaturalStatus then
         for _,s in ipairs(self._config.identity.sxbNaturalStatus) do
