--- stats/sexbound/readme.txt
@@ -1,10 +1,0 @@
-# READ ME
-
-The *_primary.lua scripts are used to override the default Starbound primary scripts.
-
-## SCRIPTS
-
-* Monster - monster_primary.lua
-* NPC     - npc_primary.lua
-* PLAYER  - player_primary.lua
-
